Separate out Docker functions from the model package
