+++ main_test.go
@@ -11,12 +11,13 @@ import (
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/client"
-	model "github.com/fubarhouse/pygmy-go/service/interface"
+	"github.com/fubarhouse/pygmy-go/service/interface/docker"
 	. "github.com/smartystreets/goconvey/convey"
 )
 
 const (
 	dindContainerName = "exampleTestContainer"
+	binaryReference = "pygmy-go-linux-x86"
 )
 
 var (
@@ -37,16 +38,16 @@ type config struct {
 // run to keep the consistency for as many tests as are required.
 func setup(t *testing.T, config *config) {
 
-	var cleanCmd = "/builds/pygmy-go-linux-x86 clean"
-	var statusCmd = "/builds/pygmy-go-linux-x86 status"
-	var upCmd = "/builds/pygmy-go-linux-x86 up"
-	var downCmd = "/builds/pygmy-go-linux-x86 down"
+	var cleanCmd = fmt.Sprintf("/builds/%v clean", binaryReference)
+	var statusCmd = fmt.Sprintf("/builds/%v status", binaryReference)
+	var upCmd = fmt.Sprintf("/builds/%v up", binaryReference)
+	var downCmd = fmt.Sprintf("/builds/%v down", binaryReference)
 
 	if config.configpath != "" {
-		cleanCmd = fmt.Sprintf("/builds/pygmy-go-linux-x86 clean --config %v", config.configpath)
-		statusCmd = fmt.Sprintf("/builds/pygmy-go-linux-x86 status --config %v", config.configpath)
-		upCmd = fmt.Sprintf("/builds/pygmy-go-linux-x86 up --config %v", config.configpath)
-		downCmd = fmt.Sprintf("/builds/pygmy-go-linux-x86 dow --config %v", config.configpath)
+		cleanCmd = fmt.Sprintf("/builds/%v clean --config %v", binaryReference, config.configpath)
+		statusCmd = fmt.Sprintf("/builds/%v status --config %v", binaryReference, config.configpath)
+		upCmd = fmt.Sprintf("/builds/%v up --config %v", binaryReference, config.configpath)
+		downCmd = fmt.Sprintf("/builds/%v dow --config %v", binaryReference, config.configpath)
 	}
 
 	Convey("Pygmy Application Test: "+config.name, t, func() {
@@ -61,7 +62,7 @@ func setup(t *testing.T, config *config) {
 			})
 
 			Convey("Image pulled", func() {
-				_, e := model.DockerPull("library/docker:dind")
+				_, e := docker.DockerPull("library/docker:dind")
 				So(e, ShouldBeNil)
 			})
 
@@ -92,7 +93,7 @@ func setup(t *testing.T, config *config) {
 		Convey("Populating Daemon", func() {
 
 			Convey("Container has started the daemon", func() {
-				_, e := model.DockerExec(dindContainerName, "dockerd")
+				_, e := docker.DockerExec(dindContainerName, "dockerd")
 				So(e, ShouldEqual, nil)
 				time.Sleep(time.Second * 2)
 			})
@@ -104,7 +105,7 @@ func setup(t *testing.T, config *config) {
 
 			for _, image := range config.images {
 				Convey("Pulling "+image, func() {
-					_, e := model.DockerExec(dindContainerName, "docker pull "+image)
+					_, e := docker.DockerExec(dindContainerName, "docker pull "+image)
 					time.Sleep(time.Second * 2)
 					So(e, ShouldBeNil)
 				})
@@ -114,7 +115,7 @@ func setup(t *testing.T, config *config) {
 		Convey("Application Tests", func() {
 
 			Convey("Container has configuration file ("+config.configpath+")", func() {
-				d, _ := model.DockerExec(dindContainerName, "stat "+config.configpath)
+				d, _ := docker.DockerExec(dindContainerName, "stat "+config.configpath)
 				if config.configpath == "" {
 					SkipSo(string(d), ShouldContainSubstring, config.configpath)
 				} else {
@@ -123,30 +124,30 @@ func setup(t *testing.T, config *config) {
 			})
 
 			Convey("Container has compiled binary from host", func() {
-				d, _ := model.DockerExec(dindContainerName, "stat /builds/pygmy-go-linux-x86")
-				So(string(d), ShouldContainSubstring, "/builds/pygmy-go-linux-x86")
+				d, _ := docker.DockerExec(dindContainerName, fmt.Sprintf("stat /builds/%v", binaryReference))
+				So(string(d), ShouldContainSubstring, fmt.Sprintf("/builds/%v", binaryReference))
 			})
 
-			d, _ := model.DockerExec(dindContainerName, "/builds/pygmy-go-linux-x86")
+			d, _ := docker.DockerExec(dindContainerName, fmt.Sprintf("/builds/%v", binaryReference))
 			Convey("Container can run pygmy", func() {
 				So(string(d), ShouldContainSubstring, "local containers for local development")
 			})
 
 			// While it's safe, we should clean the environment.
-			_, e := model.DockerExec(dindContainerName, cleanCmd)
+			_, e := docker.DockerExec(dindContainerName, cleanCmd)
 			if e != nil {
 				fmt.Println(e)
 			}
 
 			Convey("Default ports are not allocated", func() {
-				g, _ := model.DockerExec(dindContainerName, statusCmd)
+				g, _ := docker.DockerExec(dindContainerName, statusCmd)
 				for _, service := range config.servicewithports {
 					So(string(g), ShouldContainSubstring, service+" is able to start")
 				}
 			})
 
 			Convey("Pygmy started", func() {
-				d, _ = model.DockerExec(dindContainerName, upCmd)
+				d, _ = docker.DockerExec(dindContainerName, upCmd)
 				if config.configpath != "" {
 					So(string(d), ShouldContainSubstring, "Using config file: "+config.configpath)
 				}
@@ -156,7 +157,7 @@ func setup(t *testing.T, config *config) {
 			})
 
 			Convey("Endpoints are serving", func() {
-				d, _ = model.DockerExec(dindContainerName, statusCmd)
+				d, _ = docker.DockerExec(dindContainerName, statusCmd)
 				for _, endpoint := range config.endpoints {
 					if config.skipendpointchecks {
 						SkipSo(string(d), ShouldNotContainSubstring, "! "+endpoint)
@@ -170,11 +171,11 @@ func setup(t *testing.T, config *config) {
 		Convey("Environment Cleanup", func() {
 			Convey("Pygmy has cleaned the environment", func() {
 
-				_, e := model.DockerExec(dindContainerName, downCmd)
+				_, e := docker.DockerExec(dindContainerName, downCmd)
 				So(e, ShouldBeNil)
-				_, e = model.DockerExec(dindContainerName, cleanCmd)
+				_, e = docker.DockerExec(dindContainerName, cleanCmd)
 				So(e, ShouldBeNil)
-				d, _ := model.DockerExec(dindContainerName, statusCmd)
+				d, _ := docker.DockerExec(dindContainerName, statusCmd)
 				for _, service := range config.services {
 					So(string(d), ShouldContainSubstring, service+" is not running")
 				}
