+++ service/library/up.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/fubarhouse/pygmy-go/service/endpoint"
-	model "github.com/fubarhouse/pygmy-go/service/interface"
+	"github.com/fubarhouse/pygmy-go/service/interface/docker"
 )
 
 // Up will bring Pygmy up.
@@ -29,8 +29,8 @@ func Up(c Config) {
 	}
 
 	for _, volume := range c.Volumes {
-		if s, _ := model.DockerVolumeExists(volume); !s {
-			_, err := model.DockerVolumeCreate(volume)
+		if s, _ := docker.DockerVolumeExists(volume); !s {
+			_, err := docker.DockerVolumeCreate(volume)
 			if err == nil {
 				fmt.Printf("Created volume %v\n", volume.Name)
 			} else {
@@ -56,7 +56,7 @@ func Up(c Config) {
 
 			// Here we will immitate the docker command by
 			// pulling the image if it's not in the daemon.
-			images, _ := model.DockerImageList()
+			images, _ := docker.DockerImageList()
 			imageFound := false
 			for _, image := range images {
 				for _, digest := range image.RepoDigests {
@@ -71,7 +71,7 @@ func Up(c Config) {
 			// When running 'docker run', it will pull the image.
 			// For UX it makes sense we do this here.
 			if !imageFound {
-				if _, err := model.DockerPull(service.Config.Image); err != nil {
+				if _, err := docker.DockerPull(service.Config.Image); err != nil {
 					continue
 				}
 			}
@@ -91,7 +91,7 @@ func Up(c Config) {
 	// Docker network(s) creation
 	for _, Network := range c.Networks {
 		if Network.Name != "" {
-			netVal, _ := model.DockerNetworkStatus(Network.Name)
+			netVal, _ := docker.DockerNetworkStatus(Network.Name)
 			if !netVal {
 				if err := NetworkCreate(Network); err == nil {
 					fmt.Printf("Successfully created network %v\n", Network.Name)
@@ -108,7 +108,7 @@ func Up(c Config) {
 		name, nameErr := service.GetFieldString("name")
 		// If the network is configured at the container level, connect it.
 		if Network, _ := service.GetFieldString("network"); Network != "" && nameErr == nil {
-			if s, _ := model.DockerNetworkConnected(Network, name); !s {
+			if s, _ := docker.DockerNetworkConnected(Network, name); !s {
 				if s := NetworkConnect(Network, name); s == nil {
 					fmt.Printf("Successfully connected %v to %v\n", name, Network)
 				} else {
